requirements.txt
pydantic
email-validator
shapely
protobuf
jwt
oci-cli
requests
fastapi
uvicorn
sqlalchemy
psycopg2-binary
geoalchemy2
python-dotenv
